software/base-station/pyproject.toml
[tool.poetry]
name = "base-station"
version = "0.2.0"
description = ""
authors = ["Sam Korn <korn94sam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pillow = "^10.2.0"
python-multipart = "^0.0.9"
apscheduler = "^3.10.4"
jinja2 = "^3.1.3"
supervisor = "^4.2.5"
ishelly = "^0.1.4"
simple-pid = "^2.0.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
